GelError is now size_t.

+++ src/GEL2/Style/GelError.h
@@ -4,7 +4,7 @@
 #ifndef __GEL_UTIL_GELERROR_H__
 #define __GEL_UTIL_GELERROR_H__
 // - ------------------------------------------------------------------------------------------ - //
-typedef int GelError;	// An Error Code //
+typedef size_t GelError;	// An Error Code //
 
 enum {
 	GEL_OK = 0,
